Feedback Forms Prototype for QC

+++ clasp-proj/40_feedbackMain.js
@@ -3,16 +3,16 @@
 
 // --------------- This is the Main Scoring Process Caller ---------------- //
 
-function createFeedbackForms(useStepsSubset, useIndicatorSubset, thisCompany, filenamePrefix, filenameSuffix, mainSheetMode) {
+function createFeedbackForms(useIndicatorSubset, thisCompany, filenamePrefix, filenameSuffix, mainSheetMode) {
     // importing the JSON objects which contain the parameters
     // Refactored to fetching from Google Drive
 
-    var Config = centralConfig // var Config = importLocalJSON("Config")
-    var CompanyObj = thisCompany // TODO this a JSON Obj now; adapt in scope
+    var Config = centralConfig
+    var CompanyObj = thisCompany
     var IndicatorsObj = indicatorsVector
     var ResearchStepsObj = researchStepsVector
 
-    var sheetModeID = "SC"
+    var sheetModeID = "FB"
 
     var companyFilename
     if (CompanyObj.label.altFilename) {
@@ -21,11 +21,9 @@ function createFeedbackForms(useStepsSubset, useIndicatorSubset, thisCompany, fi
         companyFilename = CompanyObj.label.current
     }
 
-    Logger.log('begin main Scoring for ' + companyFilename)
     Logger.log("creating " + mainSheetMode + ' Spreadsheet for ' + companyFilename)
 
     var hasOpCom = CompanyObj.hasOpCom
-    Logger.log(companyFilename + " opCom? - " + hasOpCom)
 
     // define File name
     var spreadsheetName = spreadSheetFileName(filenamePrefix, mainSheetMode, companyFilename, filenameSuffix)
@@ -34,7 +32,6 @@ function createFeedbackForms(useStepsSubset, useIndicatorSubset, thisCompany, fi
     var File = connectToSpreadsheetByName(spreadsheetName)
     var fileID = File.getId()
 
-    // creates Outcome  page
     // TODO: redundant --> refactor to function 
     // if an empty Sheet exists, track and delete later
     var emptySheet = File.getSheetByName("Sheet1")
@@ -46,34 +43,61 @@ function createFeedbackForms(useStepsSubset, useIndicatorSubset, thisCompany, fi
         hasEmptySheet = false
     }
 
-    // Scoring Scheme / Validation
-    // TODO Refactor to module and values to i.e. Config.JSON
-    var pointsSheet
-    // var pointsSheet = insertSheetIfNotExist(File, "Points", false)
-    // if (pointsSheet !== null) {
-    //     fillPointsSheet(pointsSheet, "Points")
-    // }
-
-    // --- // Main Procedure // --- //
+    // --- // Feedback Parameters // --- //
 
     var integrateOutputs = false
     var isPilotMode = false
-    var outputParams = Config.integrateOutputsArray.feedbackParams
+    var outputParams = Config.feedbackParams
+    var subStepNr = outputParams.subStepNr
+    var hasFullScores = outputParams.hasFullScores
+    var includeSources = outputParams.includeSources
+    var includeNames = outputParams.includeNames
+    var includeResults = outputParams.includeResults
+    var dataColWidth = outputParams.dataColWidth
 
+    // var to estimate max sheet width in terms of columns based on whether G has subcomponents. This is needed for formatting the whole sheet at end of script. More performant than using getLastCol() esp. when executed per Sheet (think 45 indicators)
+    var globalNrOfComponents = 1
+    if (IndicatorsObj.indicatorClasses[0].components) {
+        globalNrOfComponents = IndicatorsObj.indicatorClasses[0].components.length
+    }
 
-    addSetOfScoringSteps(File, sheetModeID, Config, IndicatorsObj, ResearchStepsObj, CompanyObj, hasOpCom, useIndicatorSubset, integrateOutputs, outputParams, isPilotMode)
+    var companyCols = 1
+    if (hasOpCom) { companyCols = 2 }
+    var numberOfColumns = (CompanyObj.numberOfServices + companyCols) * globalNrOfComponents
+
+    // minus for logical -> index
+    var firstScoringStep = outputParams.firstStepNr - 1
+    var thisMainStep = ResearchStepsObj.researchSteps[firstScoringStep]
+
+    var thisSubStep = thisMainStep.substeps[subStepNr]
+    var thisSubStepID = thisSubStep.subStepID
+    var thisSubStepLabel = thisSubStep.labelShort
+
+    // --- // MAIN: create Single Indicator Class Sheet // --- // 
+
+    var thisIndClass
+    var sheetName
+    var lastCol
+    var blocks = 1
+
+    var doOverwrite = true
+    sheetName = outputParams.sourcesSheetname
+
+    var sourcesSheet = importSourcesSheet(File, sheetName, CompanyObj, doOverwrite)
+
+    for (var c = 0; c < IndicatorsObj.indicatorClasses.length; c++) {
+
+        lastCol = 1
+
+        thisIndClass = IndicatorsObj.indicatorClasses[c]
+
+        sheetName = thisIndClass.labelLong
+
+        lastCol = insertFeedbackSheet(File, sheetName, lastCol, isPilotMode, hasFullScores, thisIndClass, sheetModeID, thisMainStep, CompanyObj, numberOfColumns, hasOpCom, blocks, dataColWidth, integrateOutputs, useIndicatorSubset, includeSources, includeNames, includeResults, thisSubStep, thisSubStepID, thisSubStepLabel)
+    }
 
     // if existing, remove first empty sheet
-    if (hasEmptySheet) {
-        File.deleteSheet(emptySheet)
-    }
-
-    if (pointsSheet) {
-        moveSheetToPos(File, pointsSheet,1)
-        pointsSheet.hideSheet() // hide points - only possible after a 2nd sheet exists
-    }
-
-
+    if (hasEmptySheet) {File.deleteSheet(emptySheet)}
 
     return fileID
 }
